requirements.txt
flask==0.12.3
flask_restful==0.3.6
flask_script==2.0.6
flask_migrate==2.1.1
marshmallow==2.15.1
flask_sqlalchemy==2.5.1
flask_marshmallow==0.8.0
marshmallow-sqlalchemy==0.13.2
psycopg2==2.9.3
Flask-Cors==3.0.10
PyMySQL==1.0.2
Flask-WTF==1.0.0
Flask-Table==0.5.0
WTForms-SQLAlchemy==0.3